apimll/requirements.txt
django==3.0.3
scikit-learn==0.22.2
djangorestframework==3.11.1